refactor: virtualized table wrapper
