+++ packages/DataTable/src/components/VirtualizedTable/index.js
@@ -1,0 +1,1 @@
+export { default } from "./VirtualizedTable";
